refactor: 결과 카드 타입을 model 레이어로 이동
