+++ src/features/test-result/ui/ResultTabContent.tsx
@@ -1,147 +1,13 @@
 import { Button, ListRow } from "@toss/tds-mobile";
 import { adaptive } from "@toss/tds-colors";
-import { MultipleResultCard, type MultipleResultOption } from "./MultipleResultCard";
+import { MultipleResultCard } from "./MultipleResultCard";
 import { SubjectiveResultCard } from "./SubjectiveResultCard";
-import { ScaleResultCard, type ScoreBar } from "./ScaleResultCard";
-import { AbResultCard, type AbOption } from "./AbResultCard";
-import { CardSortResultCard, type CardSortCategory } from "./CardSortResultCard";
-import { TreeResultCard, type TreeResultPath } from "./TreeResultCard";
-import { FiveSecResultCard, type FiveSecAnswer } from "./FiveSecResultCard";
-
-type QuestionResult =
-  | { type: "multiple"; title: string; imageUrl?: string; options: MultipleResultOption[] }
-  | { type: "subjective"; title: string; answers: string[] }
-  | { type: "scale"; title: string; average: number; scores: ScoreBar[] }
-  | { type: "ab"; title: string; options: AbOption[] }
-  | { type: "cardSort"; title: string; categories: CardSortCategory[] }
-  | { type: "tree"; title: string; imageUrl?: string; paths: TreeResultPath[] }
-  | { type: "fiveSec"; title: string; imageUrl?: string; answers: FiveSecAnswer[] };
-
-const MOCK_RESULTS: QuestionResult[] = [
-  {
-    type: "multiple",
-    title: "전반적인 사용 경험에 만족하시나요?",
-    options: [
-      { label: "프로필", count: 32, percentage: 58, isHighlight: true },
-      { label: "설정", count: 15, percentage: 27, isHighlight: false },
-      { label: "홈", count: 8, percentage: 15, isHighlight: false },
-    ],
-  },
-  {
-    type: "subjective",
-    title: "전반적인 사용 경험에 만족하시나요?",
-    answers: Array(15).fill("만족해요"),
-  },
-  {
-    type: "scale",
-    title: "전반적인 사용 경험에 만족하시나요?",
-    average: 3.5,
-    scores: [
-      { label: "1점", height: 20, isHighlight: false },
-      { label: "2점", height: 45, isHighlight: false },
-      { label: "3점", height: 115, isHighlight: true },
-      { label: "4점", height: 35, isHighlight: false },
-      { label: "5점", height: 15, isHighlight: false },
-    ],
-  },
-  {
-    type: "ab",
-    title: "전반적인 사용 경험에 만족하시나요?",
-    options: [
-      { label: "A안 (27.2%)", height: 45, isHighlight: false },
-      { label: "B안 (62.8%)", height: 115, isHighlight: true },
-    ],
-  },
-  {
-    type: "cardSort",
-    title: "전반적인 사용 경험에 만족하시나요?",
-    categories: [
-      {
-        name: "상의",
-        items: [
-          {
-            rank: "icon-step-1-mono",
-            label: "티셔츠",
-            count: 32,
-            percentage: 58,
-            isHighlight: true,
-          },
-          {
-            rank: "icon-step-2-mono",
-            label: "반팔티",
-            count: 15,
-            percentage: 27,
-            isHighlight: false,
-          },
-          {
-            rank: "icon-step-3-mono",
-            label: "크롭티",
-            count: 8,
-            percentage: 15,
-            isHighlight: false,
-          },
-        ],
-      },
-      {
-        name: "하의",
-        items: [
-          {
-            rank: "icon-step-1-mono",
-            label: "청바지",
-            count: 32,
-            percentage: 58,
-            isHighlight: true,
-          },
-          {
-            rank: "icon-step-2-mono",
-            label: "반바지",
-            count: 15,
-            percentage: 27,
-            isHighlight: false,
-          },
-        ],
-      },
-      {
-        name: "신발",
-        items: [
-          {
-            rank: "icon-step-1-mono",
-            label: "운동화",
-            count: 32,
-            percentage: 58,
-            isHighlight: true,
-          },
-          {
-            rank: "icon-step-2-mono",
-            label: "슬리퍼",
-            count: 15,
-            percentage: 27,
-            isHighlight: false,
-          },
-        ],
-      },
-    ],
-  },
-  {
-    type: "tree",
-    title: "내 프로필을 수정하려면 어디로 가야 할까요?",
-    paths: [
-      { label: "프로필", count: 32, percentage: 58, isHighlight: true },
-      { label: "설정", count: 15, percentage: 27, isHighlight: false },
-      { label: "홈", count: 8, percentage: 15, isHighlight: false },
-    ],
-  },
-  {
-    type: "fiveSec",
-    title: "5초동안 보고 기억에 남는것은?",
-    imageUrl: undefined,
-    answers: [
-      { label: "프로필", count: 32, percentage: 58, isHighlight: true },
-      { label: "설정", count: 15, percentage: 27, isHighlight: false },
-      { label: "홈", count: 8, percentage: 15, isHighlight: false },
-    ],
-  },
-];
+import { ScaleResultCard } from "./ScaleResultCard";
+import { AbResultCard } from "./AbResultCard";
+import { CardSortResultCard } from "./CardSortResultCard";
+import { TreeResultCard } from "./TreeResultCard";
+import { FiveSecResultCard } from "./FiveSecResultCard";
+import { MOCK_RESULTS } from "../model/mock";
 
 export function ResultTabContent() {
   return (
